move palette to it's own module
