+++ crates/nu-cli/src/shell.rs
@@ -4,6 +4,7 @@ pub(crate) mod completer;
 pub(crate) mod filesystem_shell;
 pub(crate) mod help_shell;
 pub(crate) mod helper;
+pub(crate) mod palette;
 pub(crate) mod shell;
 pub(crate) mod shell_manager;
 pub(crate) mod value_shell;
